Split the monolithic test_ws_handler.py into 8 focused files mirroring the invocation suite layout
